requirements.txt
asn1crypto==1.3.0
astroid==2.3.3
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
confusable-homoglyphs==3.2.0
cryptography==2.8
defusedxml==0.6.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.3
django-allauth==0.41.0
django-clear-cache==0.3
django-core==1.4.1
django-rest-auth==0.9.5
django-webpack-loader==0.7.0
djangorestframework==3.11.0
drf-writable-nested==0.6.2
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
postgres==3.0.0
protobuf==3.13.0
psycopg2-binary==2.8.6
psycopg2-pool==1.1
pycparser==2.20
PyJWT==1.7.1
pylint==2.4.4
pyOpenSSL==19.1.0
PySocks==1.7.1
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.1.0
PythonTurtle==0.3.2
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
SQLAlchemy==1.3.20
sqlparse==0.3.0
static3==0.7.0
urllib3==1.25.8
Werkzeug==1.0.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.11.2
